requirements.txt
absl-py==2.3.1
appnope==0.1.4
asttokens==3.0.1
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
flatbuffers==25.12.19
fonttools==4.61.1
idna==3.11
ImageIO==2.37.2
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.46.0
matplotlib==3.10.8
matplotlib-inline==0.2.1
mediapipe==0.10.31
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
pooch==1.8.2
prompt_toolkit==3.0.52
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyMatting==1.1.14
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pyzmq==27.1.0
referencing==0.37.0
rembg==2.0.69
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.16.3
six==1.17.0
sounddevice==0.5.3
stack-data==0.6.3
tifffile==2025.12.20
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
urllib3==2.6.2
wcwidth==0.2.14
